refactor: split FileHelper into focused classes
Move pure clipart image processing (trim/normalize/margins) into a new
ClipartImageProcessor and low-level file primitives (path, write,
unique filename) into a new FileStorage, shrinking the FileHelper god
class. No behavior change.

Part of #1852.
